requirements.txt
arxiv==1.4.8
python-dotenv==1.0.0
google-generativeai>=0.4.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
PyPDF2
pytesseract
Pillow 